Updated openstack/openstack Project: openstack-infra/project-config  5dc74baaba412bf5f5ddce993fa54f5d7238016b
Update infra project ACLs

To support the changes outlined in Iea53a45bb533968a13a85c0e7fbb0787325e4095
update the ACLs for all current infra projects.  The general approach is to
create a -core and -release group for every infra project, and then, if they
share commonalities, deal with that by using group inclusion in Gerrit.  This
reduces friction when adding core members to new projects.

The intent of this change is normalize all of the relevant configs so that
the -core group is used for reviews and the -release group for tags.  No
current infra projects have stable branches, so ACLs for proposed/ branches
were removed.  Similarly branch creation permissions were removed.  The
contributor agreement flag should not be changed.

Work will need to be done in Gerrit to set up these new groups before
this change merges for a seamless transition.

Change-Id: I44750a59e377593d89538c2e590eb78303c8e413