Updated openstack/openstack Project: openstack/nova  f571899bcfafee6c39b4d84514b7cafb27426b29
tests: libvirt: Fix test_volume_snapshot_delete tests

Some of the tests remove _RELATIVE flag from fakelibvirt
module, to simulate old libvirt, which can cause other tests
to fail when they try to remove the same flag. This fail can
occur intermittently based on the order in which the tests
are invoked.

fakelibvirt already has the _RELATIVE flags defined in the
module. The tests should not update/add them and should only
remove them to simulate the absence of flag (re-adding them
at the end of testcase). This way irrespective of the order
in which the tests are run, it won't cause test to fail when
they try to remove _RELATIVE flag.

Change-Id: I231cef8727a78e60b869fdbf057072ea5baee9b4