Updated openstack/openstack Project: openstack/diskimage-builder  8efa1fe1540fcea69a9cacaac587ea9b7018bd08
rhel-common element should not use attach with activation key

In the case of using portal registration with an activation key, the
rhel-common element is still executing a `subscription-manager attach`
command. This should not happen if an activation key is provided. This
is because an activation key already provides the subscriptions to
attach.

This patch fixes this behavior.

Change-Id: I5a8425d1778362bb7a0dadc91a46308f16b2a526
Closes-Bug: #1456648