Updated openstack/openstack Project: openstack/tempest  58e735217400425d3ce0901baec9936e5de92568
Add OS_TOP_LEVEL to testr conf

When installing tempest, it could be that the root of tempest code
is not a subfolder of the folder where the testrepository
configuration is held, and where test results are stored.
unittest2 will refuse to run tests for which the discovery root
is not a subfolder of the specified top level folder.

To fully support this scenario, adding OS_TOP_LEVEL in .testr.conf
and defaulting it to "./" which is the current value.

Change-Id: I60f788051d6a10ab00aecef5d09093a68f7f4030