Updated openstack/openstack Project: openstack/neutron  cd44750e9d52ee5c846f6406a69ee4f4a8d7cb52
py34: don't run any tests except unit tests

py34 job was intended for unit tests only. It's important to distinguish
between different types of tests, because they all have different
requirements to execution environment. E.g. functional tests are not
expected to run in a restricted env designed for unit tests, and that's
even more valid for fullstack tests.

Otherwise, the job may fail or apply irrecoverable changes to test
runner system, breaking the system.

If we ever want to support py3 for other types of tests, we should add
separate jobs just for that.

Note that the neutron-python3 blueprint was not intended to introduce
changes to support anything but unit test execution with the new Python
version, so strictly speaking, any effort to make other test types work
is out of scope.

Change-Id: Ia96f03f05269c1938b51ee26f4b4075b69967bb8