Updated openstack/openstack Project: openstack/nova  9aa11680d52d7959c1ba35d33752ab2f3fc51cbf
Create instance_extra entry if it doesn't update

When updating an instance, if there is no instance_extra, for older
instances, we make sure we create one if there is not one already.

This keeps flavor data migrations (ie system_metadata ->
instance_extra) from failing when the system_metadata entries are soft
deleted but an instance_extra entry is not created because it was
"updated" assuming an instance_extra entry already existed, leaving
the instance with no flavor data.

Closes-Bug: #1459758
Change-Id: I715217d98729c6c5ac45d64d8bacaaca745ff3d7