Updated openstack/openstack Project: openstack/nova  3eca7f3af18b5cf7a90749d4b01f2a724f056dfe
Hyper-V: Fixes cold migration / resize issue

Commit 3e42c7ae3f7a9353834c8c9c4dc79ee39dd0783b changed where the Hyper-V
VM configuration files are stored. The files are being stored in the same
folder as the instance. Performing a cold resize / migration will cause a
os.rename call on the instance's folder, which fails as long as there are
configuration files used by Hyper-V in that folder, thus resulting in a
failed migration and the instance ending up in ERROR state.

Hyper-V configuration files are stored in a subfolder of the instance's
folder (e.g.: C:\OpenStack\Instances\instance-000000b6\Virtual Machines).

This commit makes sure that Hyper-V configuration files are not being
moved during migration by skipping subfolders in the instance's folder.

Change-Id: I4874eed7d284fab241e4371e550a88270ca156f1
Closes-Bug: #1453835