Updated openstack/openstack Project: openstack-infra/release-tools  fe1787333a754e7c3de96b05c2959b74113f972c
Include HEAD SHA in unreleased log output

Include the SHA of the current HEAD commit on the branch being queried
so we can copy that value to build a script to release the versions
reported. This lets us produce a proposed list of releases, ask the PTLs
to verify them, and then release exactly what we discussed even if it is
a couple of days later and more patches have merged.

Change-Id: I0720b88fc2e8f36310f3b585ff8f89640a8f8b39