Updated openstack/openstack Project: openstack/cinder-specs  6ec3788b8241c595bf686dcc4143cc95345492e2
Non Disruptive Backup

Currently a backup operation can only be performed when the volume is
detached. This blueprint proposes the following changes in order to do
a non-disruptive backup while the volume is still attached:

For the LVM driver, a temporary snapshot will be created and backup
will be performed by reading from the local path of the snapshot.
After that, the temporary snapshot will be deleted.

For other drivers, a default implementation will be provided to
create a temporary volume from the source volume and backup will be
performed by attaching the temporary volume and reading from it.
After it is done, the temporary volume will be detached and deleted.

For other drivers, a more optimized approach will also be provided to
create a temporary snapshot, attach the snapshot, backup the snapshot,
and then detach and delete the temporary snapshot. This optimized
approach is not required. If a driver has not implemented attach
snapshot, the default implementation with a temporary volume will
be used.

DocImpact
blueprint non-disruptive-backup

Change-Id: I069d86d315dd2c83d43d694087f2dcc2f0b0752b