Updated openstack/openstack Project: openstack/requirements  8cb51513a92f81f3bf46bf4b7e6ff1439dbd8817
Bump ddt to min version 0.7.0

Manila and manilaclient projects use lib "ddt" for testing. Both use feature
of this lib that appeared only starting with version 0.7.0 [1]
It worked till that moment because latest version was installed each time.
But it can cause errors in theory if current version 0.4.0 will be installed
indeed.

[1]: https://github.com/txels/ddt/commit/c55bbf5d

Change-Id: Iea6881a5efc7e3203747603727bec51bf4a4c0d8