Updated openstack/openstack Project: openstack/keystonemiddleware  2d4e19404aa33200767dab86956608878876b03a
Ensure cache keys are a known/fixed length

Do not assume a token_id will result in a sane length for a memcache
key length. In cases such as Fernet, these ids can easily exceed the
limit on memcache key size. This change ensures we always use a SHA256
of the token id passed in, resulting in a fixed length cache key.

Change-Id: I550e0a1b190047438756bbf40490815a5f177ea7
Closes-Bug: #1460225