Updated openstack/openstack Project: openstack-infra/project-config  f1b793bbb97a4a901545eb141a408f1a74b0538a
Expand the vote range for governance

The voting configuration for governance produces a red-x and
green-check for +/-1 code reviews now, which is confusing to people
since they normally associate those icons with +/-2 votes.  This
change expands the vote range to include +/-2 but gives no one
permission to vote at those levels.  This should cause Code-Review
votes to appear as +/-1, while the NoOp function continues to
mean that +2 votes are not necessary.

Change-Id: I31adfe3c847d181809825e1030e8c165e67ae9aa