Updated openstack/openstack Project: openstack-infra/jeepyb  bee95e5e6318839d868dd0c4d9b451fbe3daa910
Updated code review timeline

The current welcome message for new contributors mentions that it usually
takes 3 to 7 days for the code to be reviewed.  But the code also needs
approval from the core member(s) for the merge to happen.  It generally
takes at least 2 weeks for core to review the changes as they are busy
with other high priority things.  For a newbie it is not very clear on
how long the patch will take to get reviewed and merged.  Hence added
more text related to this.

Change-Id: I78d32a99338b74c648f68d87a5c7bcb577e016b0