Updated openstack/openstack Project: openstack/requirements  dbfab46a65d355f07ddf418957768659861f44f5
Add windows requirements.

These are existing requireements of Neutron.  They were previously
installed by a build hook.  That caused problems so they were moved to
being specified in setup.py.  Now they need to be moved to
requirements.txt so that setup.py can be kept in sync with the global
version.

Change-Id: I2369971d306c10dc39a1b89698cec95cf7551d07