Updated openstack/openstack Project: openstack-infra/system-config  447e75fa84900c674b2fbd43aa0374eb22eeca33
Generate machine-readable tree of change approvers

* tools/who-approves.py: New script which, when run using
OpenStack's Gerrit server, builds JSON and YAML representations of
repos with information on the official owning project if any,
integration status, and groups with approve rights listing the
members of each along with their Gerrit preferred E-mail addresses
and usernames when available.

It was done as a demonstration to a representative of a foundation
member company who requested a list of the "core reviewers" for
official projects, optionally broken down by integrated vs. other.
I'm attempting to show that this data is already publicly available
and can be extracted/analyzed by anyone without needing to request
it.

Change-Id: I017d98e2a45d5082fce8f4eecb5e6f8633c02ac8