Updated openstack/openstack Project: openstack/nova  30564b9207cdc5b93a971780f88399996ff0b876
Force the value of LC_ALL to be en_US.UTF-8

On Python 3, if LC_ALL is not passed then the interpreter will assume it
is C (which causes the interpreter to use ASCII). This, in turn, causes
the unicode in the README.rst for rfc3986 to trigger a UnicodeDecodeError
exception.

jlvillal (John Villalovos) pointed out
I340c88ea7d098ca5943d60adc73f63a0af79a405 in IRC as the fix to this bug.

Closes-bug: 1460206
Change-Id: If05e3fe76450c06431413f63e46e154cf44fc1cf