Updated openstack/openstack Project: openstack-infra/project-config  13d4ecff60f54c7b2f04e9f0dbb6578c2feaf774
Introduce an experimental job for Linux Bridge

The Linux Bridge ML2 driver for Neutron needs to be tested
to ensure that changes to Neutron do not break this mechanism driver.
Operators that are not interested in SDN solutions but still interested
in using Neutron will want to use the Linux Bridge driver for their
deployments, so we should test it.

Change-Id: Iccba951301d2c09e01b6c1a15ed1e88c8e0db69e