Updated openstack/openstack Project: openstack-infra/zuul  7fca9c1cc625dec94b1c06a6a65216cd1a041e85
Allow 'MAY' in Gerrit submit records

If a review label uses NoOp, then it will indicate 'MAY' as the
status in the submit record for that label (indicating that there
are no criteria for that label to be satisfied).  Treat that in
the same way as OK so that a merge can occur.

Change-Id: I24be23d16aa133a3323b79bd775d42678ebc1dfe