Updated openstack/openstack Project: openstack/python-cinderclient  2f97e1e3ca070d9d2450519f5ebcf9c0d624cb65
Bump hacking to >=0.10.0,<0.11 to fix failure of gate pep8

pbr 1.0.1 released in 2015-05-19 requires the newer
hacking module >=0.10.0,<0.11 than the versions
specified in the test-requirement.txt file,
which causes failure in gate-python-cinderclient-pep8.

This patch bumps the pbr version in test-requirement.txt
to make it consistent with requirements repository,
and adds some fixups to avoid some new hacking rules.

Change-Id: I7de6e04e30a475ec1c26224461faf0c708178a40
Closes-Bug: #1460774