Updated openstack/openstack Project: openstack/keystone  2bb8a3cb11c396fb7404168e63a77f2db930499c
Log info for Fernet tokens over 255 chars

Fernet tokens were designed to be short and maintain some of the usability of
UUID-like tokens by being copy-pastable. Depending on the deployment
configuration, we could still generate tokens that are over 255 characters. If
this is the case, we should log some information.

Change-Id: I3aa52f7e09c4cf717ddd28d14ee36b80f53b012b