Updated openstack/openstack Project: openstack/cinder  69452c40c9f011b0f2085fee751f37e935ce0481
NetApp FC driver shims missing manage/unmanage

The NetApp FC drivers' passthrough shim layers are missing the
manage/unmanage methods, causing the manage/unmanage functions
to fail.  The necessary functionality is already present in
the underlying library layer, so the trivial fix is to add the
functions to the passthrough layers. This patch also adds a
unit test to ensure the four block storage driver interfaces
remain in sync.

Change-Id: I0fa7e16b23e5148f82a8b51bf4380f4384c35759
Closes-Bug: #1444517