Updated openstack/openstack Project: openstack/horizon  99a3ebeda4fd7d9e8673b34049206293d6d19445
Instance overview renders erroneous markup

Under Instance Detail | Overview | Meta | Key Name an unclosed </em>
is rendered for an instance launched without an associated key pair.
This seems to be reproducible using Django 1.6.1 in local devstack.
The issue seems to be that Django's implementation of the add filter
has an issue concatenating a string and a function to localize text.
Since that fails and returns an empty string which subsequently is
added to the closing </em> tag, the closing tag is the only thing
that is displayed.

This fix saves the localized value to a variable which is then used
in the add filter.

Closes-Bug: #1423797
Change-Id: I4452ca4a902bbaf19ae5b2bd9bcbe2900069ac3b