Updated openstack/openstack Project: openstack/requirements  cad2e865b793e4245e0e820942ce18b976a6c256
Add InfluxDB

Gnocchi will get a new storage driver: InfluxDB

This is python module to use the InfluxDB API.

gnocchi driver: https://review.openstack.org/#/c/165407/

The library is released in MIT License.

The library is released ~ 2 times by month.
https://github.com/influxdb/influxdb-python/releases

The source code is automatically tested by the travis-CI.
https://travis-ci.org/influxdb/influxdb-python

And the code is well covered.
https://coveralls.io/r/influxdb/influxdb-python

The documentation is autogenerated:
https://influxdb-python.readthedocs.org/en/latest/

The library is python3 compatible.

This is already packaged in ubuntu vivied and debian jessie (but outdated)
Fedora have packages for the server, and the python binding
will arrived very soon with the 2.2.0 version.

Also, upstream maintains its own packages for all
distributions, too.

This version have been choosen because of recently bug fixed
for gnocchi.

Change-Id: Ifa1da4e3c7b1372c9a2aeb021542284e5a139dcf