Updated openstack/openstack Project: openstack-infra/project-config  43b100c42ef8d6f5503f46d40eca52f35459868c
Don't use set +e to check the return value

In propose_update, we were pushing the shell's set -e functionality
out of the way, so we could use a if block to check the return value.
We only care if the command exits non-zero, so we can do it in an if
directly.

Change-Id: I196751e79aee24e96a854fec4a51b4756c96047f