Updated openstack/openstack Project: openstack/neutron-lbaas  fa38f1aaa7a22272a811ea3e030a5897875d3815
Pull lbaasv1 tempest tests in-tree, use installed neutron for v1/v2 job

Fix the lbaas v2 api job to use the installed neutron and neutron-lbaas
in CI land. I'm not super happy about tox still being in the mix, so
expect it to run testr directly in a future commit, but the output
gets all horked when I switch, and I'd rather get the job fixed and
tackle that second.

Also pulls in a subset of tempest, to be used during the migration
to tempest-lib. Future test commits need to move us away from the
in-tree tempest, and start removing it.

Also yank the check_bash check, which was removed from neutron.

Change-Id: I2790239b8cd361bc4ac7905b4118f34648b97840