Updated openstack/openstack Project: openstack/keystone  f6c01dd1673b290578e9fff063e27104412ffeda
Adds inherited column to RoleAssignment PK

It should be possible to add both inherited and
non-inherited role assignments for the same actor
and target with the same role.

However, this was not currently possible since
the inherited column was not part of the PK of
the assignment table.

This patch alters the table definition and adds a
migration script and tests for it.

Closes-Bug: #1403539

Change-Id: I1ba4935934b0dc6b6077d18761023ad50462c8b8