Updated openstack/openstack Project: openstack-infra/release-tools  8572d2df36cf7b4217ddb5a91fa4085aefcaf201
improve email format with stable and other tags

Separate the series specification from the stable branch flag so we can
always include a series in release notes subjects.

Add an option to specify other email topic tags to be added to the
beginning of the subject line, so we can add project names for libraries
where the project may not be obvious.

Change-Id: I8dd662c24257346cee76128643b4eee202a89dd7