Updated openstack/openstack Project: openstack/requirements  f7f0d160362fe542d94c00a2bd6308b2e60d9fc9
Add sqlalchemy-utils to requirements

This library is actively maintained, last release is 5 days old.
This library is good code.
This library is BSD licensed.
This library is (being) packaged (already used by Gnocchi).
This library function is not covered by other libraries.
This library is going to be used by Ceilometer.

Change-Id: I1bd9e9cde33c08d6a7efd69ec2efec342d3487fd