Updated openstack/openstack Project: openstack-infra/project-config  eea0ab7d0418ecb10465625cacbea334264210fb
Py26 for Keystonemiddleware on Icehouse/Juno Only

Only gate the py26 gate job on Icehouse and Juno for the Keystonemiddleware
repository. Since it is not supported to run very modern Keystonemiddleware
with much older services, py26 is no longer needed for Kyestonemiddleware.

This will allow us to cleanup the test-requires and other py26-specific
requirements from the project.

Change-Id: I316d3ec56ade662cb5deeca0d3d48230c878f35d