Updated openstack/openstack Project: openstack/keystone  386054847e3434122a5070f2894908ff437d901e
Switch from deprecated isotime

oslo_utils.timeutils.isotime() is deprecated as of 1.6 so we need
to stop using it.

The deprecation message says to use datetime.datetime.isoformat()
instead, but the format of the string generated by isoformat isn't
the same as the format of the string generated by isotime. The string
is used in tokens and other public APIs and we can't change it
without potentially breaking clients.

So the workaround is to copy the current implementation from
oslo_utils.timeutils.isotime() to keystone.common.utils.isotime().

Closes-Bug: 1461251

Change-Id: Icf821be8cd4ee7f52c1a3d89bfed302488c080f2