Updated openstack/openstack Project: openstack/python-swiftclient  f0cc3be2ac2cb6474ba117da323f6369c1e6c791
Make default get_auth timeout be None

Setting timeout to a default of False in get_auth()
results in a requests timeout of 0.0 in keystoneclient,
which causes a connection failure.

This bug will cause func tests to not run when using
keystoneauth.

Added unit tests to verify correct default timeout is set
in get_auth().

Drive-by: remove what seems like a stale TODO comment

Change-Id: I17c781ce160a682b1768d315422ade0cdd2df198