Updated openstack/openstack Project: openstack/nova  e56c984d8cf6be866e3ea87bbdfd736f18766180
Add nova object equivalence based on prims

This adds a utility method called obj_equal_prims() to
nova.objects.base.py.

The utility method compares two nova objects for equivalence
by comparing their primitive representation, ignoring some
of the keys in the primitives. In particular it will ignore
the nova_object.changes keyword. A list of additional keys
to ignore can be passed in as a parameter by the caller.

This equivalence method will be needed when the resource tracker
is converted to use the ComputeNode object in a dependant patch.

Change-Id: I1b46878bd5d1172cca49b7a7bbb7cabb8a02ed11