Updated openstack/openstack Project: openstack/nova  93cbba96fa3d92ce1a853516171dcf69a3ec3984
bdm: Make sure that delete_on_termination is a boolean

Make sure that 'delete_on_termination' is always a boolean, even if it
was not passed in and thus got defaulted to None up until now.

This patch is really more of a pedantic fix, as we use objects for BDMs
everywhere outside of the API service, and object fields are of the
correct type.

Related-bug: #1370177

Change-Id: I2724bcbe159490f3bdd85f833412f3b20c4e1e23