Updated openstack/openstack Project: openstack/trove  0d5de65cc8268a8f97213e3ea0f62caac3ff9cf8
Fixes a broken debug message in schema matching

This rather useful debug tool in the bowels of common/wsgi.py produces
a misleading message because, most often, the matching_schema has no
property called 'name'. Fixed it to print the schema in its entirety
because in most cases one would take the name and go lookup the schema
anyway.

The other way around would be to go and change all of apischema.py and
add a name property. This seemed to be the less intrusive and
disruptive change.

Change-Id: I4075bc911d2cc07cb6e7d98ffcff96441f7b3416
Closes-Bug: #1460033