Updated openstack/openstack Project: openstack/horizon  93e85486ac77436848e1ff3100b084e899bea888
filter the subnets which already in the router

When we already add the subnet to router,
if we click 'Add Interface' again, we can see
the subnet also can select, but if we real to
add the subnet, it report a error. We can hidden
the subnet if it already add to the router.

Closes-bug:#1417892

Change-Id: I485608d267b9b235c1e83a990e330dd1ca98f17e